requirements.txt
streamlit>=1.28.0
ultralytics>=8.0.220
opencv-python-headless>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pandas>=2.1.0
plotly>=5.17.0
requests>=2.31.0